requirements.txt
openai==1.93.2
python-dotenv==1.1.1
psutil==7.0.0
requests==2.31.0
flask
